+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityEditModelTest.java
@@ -58,8 +58,8 @@ public class SwingEntityEditModelTest {
 	@Test
 	void comboBoxModel() {
 		FilterComboBoxModel<String> model = employeeEditModel.comboBoxModel(Employee.JOB);
-		model.includeNull().set(true);
-		model.nullItem().set("null");
+		model.items().nullItem().include().set(true);
+		model.items().nullItem().set("null");
 		assertNotNull(model);
 		assertEquals(model, employeeEditModel.comboBoxModel(Employee.JOB));
 		employeeEditModel.refreshComboBoxModels();